+++ mapi/include/TCM_values.h
@@ -6,7 +6,7 @@
 
 struct TypeTCM {
     std::map<std::string, std::string> addresses;
-    std::map<std::string, std::vector<std::vector<uint32_t>>> tcmWords;
+    std::map<std::string, std::vector<std::vector<int64_t>>> tcmWords;
     std::map<std::string, std::pair<std::string, std::string>> tcmEquations;
     struct ActualValues {
         int32_t DELAY_A         :16,  //┐
